pyproject.toml
[tool.poetry]
name = "forecasting-tools"
version = "0.2.38"
description = "AI forecasting and research tools to help humans reason about and forecast the future"
authors = ["Benjamin Wilson <mokoresearch@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = ["ai", "forecasting", "research", "metaculus", "prediction", "future", "market", "time-series"]
repository = "https://github.com/Metaculus/forecasting-tools"
packages = [{include = "forecasting_tools"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Natural Language :: English",
    "Operating System :: OS Independent"
]

[tool.poetry.scripts]
forecasting = "forecasting_tools.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.51.0"
tiktoken = ">=0.8.0,<0.10.0"
aiofiles = "^24.1.0"
aiohttp = "^3.9.3"
aiolimiter = "^1.1.0"
asyncio = "^3.4.3"
nest-asyncio = "^1.5.8"
requests = "^2.32.3"
pydantic = "^2.9.2"
python-dotenv = "^1.0.0"
regex = "^2024.11.6"
tenacity = ">=8.0.0,<10.0.0"
pillow = ">=9.0.0,<12.0.0"
transformers = "^4.44.1"
typeguard = "^4.3.0"
scikit-learn = "^1.5.2"
litellm = "^1.59.1"
numpy = ">=1.26.0,<3.0.0"
exceptiongroup = "^1.2.2"
scipy = "^1.12.0"
faker = "^37.0.1"
plotly = ">=5.24.1,<7.0.0"
pandas = "^2.2.3"
streamlit = "^1.20.0"
typing-extensions = "^4.9.0"
pytz = "^2023.3"
stqdm = "^0.0.5"
cattrs = "^23.2.3"
packaging = "^23.2"
MarkupSafe = "^2.1.3"
playwright = "^1.52.0"
statsmodels = "^0.14.2"
matplotlib = "^3.9.2"

[tool.poetry.extras]
guardrails = ["nemoguardrails"]
timeseries = ["prophet", "tensorflow"]
backtesting = ["plotly", "matplotlib", "pandas"]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
seaborn = "^0.13.2"
matplotlib = "^3.9.2"
ipykernel = "^6.29.5"
notebook = "^7.2.2"
pytest = "^8.3.2"
pytest-xdist = "^3.3.1"
pytest-mock = "^3.14.0"
pytest-asyncio = ">=0.25.0,<0.27.0"
pypistats = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
